+++ Source/WebCore/dom/ComposedTreeAncestorIterator.h
@@ -25,7 +25,6 @@
 
 #pragma once
 
-#include "ElementRareData.h"
 #include "HTMLSlotElement.h"
 #include "PseudoElement.h"
 #include "ShadowRoot.h"
