+++ Source/WebCore/dom/ComposedTreeIterator.h
@@ -26,7 +26,6 @@
 #pragma once
 
 #include "ElementAndTextDescendantIterator.h"
-#include "ElementRareData.h"
 #include "HTMLSlotElement.h"
 #include "ShadowRoot.h"
