+++ Source/WebCore/dom/ElementRareData.h
@@ -41,7 +41,6 @@
 #include "PseudoElementIdentifier.h"
 #include "RenderElement.h"
 #include "ResizeObserver.h"
-#include "ShadowRoot.h"
 #include "SpaceSplitString.h"
 #include "StylePropertyMap.h"
 #include "StylePropertyMapReadOnly.h"
@@ -70,10 +69,6 @@ public:
     unsigned childIndex() const { return m_childIndex; }
     void setChildIndex(unsigned index) { m_childIndex = index; }
     static constexpr ptrdiff_t childIndexMemoryOffset() { return OBJECT_OFFSETOF(ElementRareData, m_childIndex); }
-
-    void clearShadowRoot() { m_shadowRoot = nullptr; }
-    ShadowRoot* shadowRoot() const { return m_shadowRoot.get(); }
-    void setShadowRoot(RefPtr<ShadowRoot>&& shadowRoot) { m_shadowRoot = WTF::move(shadowRoot); }
 
     CustomElementReactionQueue* customElementReactionQueue() { return m_customElementReactionQueue.get(); }
     void setCustomElementReactionQueue(std::unique_ptr<CustomElementReactionQueue>&& queue) { m_customElementReactionQueue = WTF::move(queue); }
@@ -184,8 +179,6 @@ public:
             result.add(UseType::Dataset);
         if (m_classList)
             result.add(UseType::ClassList);
-        if (m_shadowRoot)
-            result.add(UseType::ShadowRoot);
         if (m_customElementReactionQueue)
             result.add(UseType::CustomElementReactionQueue);
         if (m_customElementDefaultARIA)
@@ -243,7 +236,6 @@ private:
     AtomString m_effectiveLang;
     const std::unique_ptr<DatasetDOMStringMap> m_dataset;
     const std::unique_ptr<DOMTokenList> m_classList;
-    RefPtr<ShadowRoot> m_shadowRoot;
     std::unique_ptr<CustomElementReactionQueue> m_customElementReactionQueue;
     std::unique_ptr<CustomElementDefaultARIA> m_customElementDefaultARIA;
     const std::unique_ptr<FormAssociatedCustomElement> m_formAssociatedCustomElement;
@@ -291,7 +283,6 @@ inline ElementRareData::ElementRareData()
 
 inline ElementRareData::~ElementRareData()
 {
-    ASSERT(!m_shadowRoot);
     ASSERT(!m_beforePseudoElement);
     ASSERT(!m_afterPseudoElement);
 }
@@ -366,24 +357,6 @@ inline ElementRareData* Element::elementRareData() const
 {
     ASSERT_WITH_SECURITY_IMPLICATION(hasRareData());
     return downcast<ElementRareData>(rareData());
-}
-
-inline ShadowRoot* Node::shadowRoot() const
-{
-    return hasShadowRoot() ? downcast<Element>(*this).shadowRoot() : nullptr;
-}
-
-inline ShadowRoot* Element::shadowRoot() const
-{
-    return hasShadowRoot() ? elementRareData()->shadowRoot() : nullptr;
-}
-
-inline void Element::removeShadowRoot()
-{
-    RefPtr shadowRoot = this->shadowRoot();
-    if (!shadowRoot) [[likely]]
-        return;
-    removeShadowRootSlow(*shadowRoot);
 }
 
 } // namespace WebCore
